first pass of changes after RFC
