+++ torchtune/modules/transforms/vision/__init__.py
@@ -1,0 +1,15 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from .get_canvas_best_fit import get_canvas_best_fit  # noqa
+from .resize_with_pad import resize_with_pad  # noqa
+from .tile_crop import tile_crop  # noqa
+
+__all__ = [
+    "get_canvas_best_fit",
+    "resize_with_pad",
+    "tile_crop",
+]
